Fixed incorrect wall textures

+++ map.h
@@ -30,7 +30,7 @@ typedef struct sector {
 
 typedef struct sidedef {
   int16_t  x_off, y_off;
-  uint16_t upper, lower, middle;
+  int      upper, lower, middle;
   uint16_t sector_idx;
 } sidedef_t;
